# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Postal Delivery Operations & Optimization System API
  version: 0.1.0
  description: >
    Admin API for postal delivery operations across multiple post offices,
    beats, postmen and deliveries. See /api/v1/* routers in source for the
    authoritative, up-to-date contract; this document covers the primary
    resources.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /imports/upload:
    post:
      summary: Upload a CSV/XLSX/PDF delivery file and receive an import preview
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
      responses:
        "201": { description: Import created with preview }
  /imports/{id}/confirm:
    post:
      summary: Confirm a previewed import and write deliveries to the database
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Import confirmed }
...